Move useful remnants from sadbox to sys-happybox
