+++ librz/cons/dietline.c
@@ -473,7 +473,7 @@ RZ_API int rz_line_hist_load(const char *file) {
 	if (!path) {
 		return false;
 	}
-	if (!(fd = rz_sandbox_fopen (path, "r"))) {
+	if (!(fd = rz_sys_fopen (path, "r"))) {
 		free (path);
 		return false;
 	}
@@ -503,7 +503,7 @@ RZ_API int rz_line_hist_save(const char *file) {
 			}
 			*p = RZ_SYS_DIR[0];
 		}
-		fd = rz_sandbox_fopen (path, "w");
+		fd = rz_sys_fopen (path, "w");
 		if (fd != NULL) {
 			if (I.history.data) {
 				for (i = 0; i < I.history.index; i++) {
