+++ librz/include/rz_cons.h
@@ -12,7 +12,6 @@ extern "C" {
 #include <rz_util/rz_log.h>
 #include <rz_util/rz_num.h>
 #include <rz_util/rz_panels.h>
-#include <rz_util/rz_sandbox.h>
 #include <rz_util/rz_signal.h>
 #include <rz_util/rz_stack.h>
 #include <rz_util/rz_str.h>
