+++ librz/util/Makefile
@@ -15,7 +15,7 @@ OBJS=binheap.o mem.o unum.o str.o hex.o file.o range.o
 OBJS+=prof.o cache.o sys.o buf.o w32-sys.o ubase64.o base85.o base91.o
 OBJS+=list.o flist.o chmod.o graph.o event.o alloc.o print_code.o
 OBJS+=regex/regcomp.o regex/regerror.o regex/regexec.o uleb128.o
-OBJS+=sandbox.o calc.o thread.o thread_sem.o thread_lock.o thread_cond.o
+OBJS+=calc.o thread.o thread_sem.o thread_lock.o thread_cond.o
 OBJS+=strpool.o bitmap.o time.o format.o pie.o print.o utype.o
 OBJS+=seven.o randomart.o zip.o debruijn.o log.o getopt.o table.o
 OBJS+=utf8.o utf16.o utf32.o strbuf.o lib.o name.o spaces.o signal.o syscmd.o
